Wipeout thread locals after each request
This is necessary as the BatchLoader gem is using Thread locals to
store the loaded values. If we don't wipeout the Thread locals after
each request, we can endup having a memory bloat and even worst these
values can leak into next request which might result in with having
inconsistent responses.