Allow SCIM API to return SAML identity in certain cases
Prior to separating SCIM and SAML identities, the SCIM API could
be used to get SAML-only users. Some customers came to rely on
this for auditing purposes. Despite it not being an intended use
of the API, we're reintroducing the functionality temporarily
while we address some additional deficiencies in SCIM support.
This feature is behind a feature flag and will only be enabled
for customers with a broken workflow.