Use a memoized Flipper engine in tests
This makes us to use a memoized Flipper engine
that simply always works.

We also change a implementation of `persisted_names`
to use a `Flipper` instead of requesting features
manually.

By default a memoized engine is used, which
can be overwritten with `stub_feature_flags: false`.

The memoized engine requires us to use object that
are `FeatureGate-inherited`. We ensure
that this holds true, always.

This MR also updates all tests to adhere to this requirement.