Add framework to allow primary user email to move to Email model
Currently the primary user email is stored as an attribute (and
associated attributes) on the User model itself. Additional user
emails are stored separately in the Email model. This change
provides a framework for combining all user emails under the Email
model. A future iteration will migrate primary emails and start
pointing to the Email model rather than attributes on the User
model.