Ensure .git/config is updated for forks
Normally when a project is created, `Projects::CreateService` attempts
to initialize a Git directory and update the `.git/config` to include
the human-readable project full path. However, for forks and project
imports, the repository is imported in a separate background task. As a
result, attempts to write to the `.git/config` directory will fail
unless the job has completed first.

To ensure we always update `.git/config` after the import is complete,
we call this update in `Project#after_import`.

Relates to https://gitlab.com/gitlab-org/gitaly/-/issues/2902