Support group wikis in git-access
This re-organizes the git-access hierarchy to be less project-centric,
in anticipation of being fully polymorphic over `HasRepository`
containers. This is an incremental step, where we first use `Project`,
`Group` and `Snippet` as possible containers, but we aim to move to a
mapping between repository type and container:

| Repo-Type | Container Model |
|-----------|-----------------|
| Code      | `Project`       |
| Wiki      | `Wiki` (as `ProjectWiki` or `GroupWiki`) |
| Design    | `DesignCollection` |
| Snippet   | `Snippet` (as either `Snippet` or `PersonalSnippet`) |

Such a move will mean some more thorough changes to the current design,
so this is just a first step and includes such temporary measures as
explicit predicates for the kind of container we expect to be using,
explicitly passing different ability names and not-implementing all of
an interface when passing an argument (`Group` does not implement all of
the methods that can be called in `UserAccess`, but currently only one
is called).