Design-Management: Update types in our graphql queries
This changes ~frontend queries used in our clients so that they refer to
the new specific ID types corresponding to the GlobalID for the domain
object in question. These changes are necessary since types must match
exactly.

No changes to the the actual front-end code itself is needed, since the
representation of these values has not changed.