Seperate class for project-specific User-access
This moves project specific checks from `UserAccess` into a new class,
`UserAccessProject`, since at all places in our code we know statically
if we need the methods that are project-specific.

The advantage of this is that we don't have to implement methods for
classes that don't make sense for them, and this opens a path to
different kinds of containers in the future.