Allow limited scalar inheritance
To smooth over the backwards incompatible nature of the global-ID
changes, this commit introduces a compatibilty layer, which allows
variables of type `ID` to be passed to arguments of type `GlobalID` (or
any specific sub-type).

In order for this to work, the use-site must be careful to coerce the
argument manually. The manual coercion can be removed when the
compatibilty layer is removed following the deprecation cycle.