Add regression test for `each_batch`
This test case demonstrates the race condition problem of the current
implementation of `each_batch` query method.

This case may look unlikely to happen to readers but the real problem
here is not creating records in database in the very last step of the
iteration. The real problem is transaction visibility feature of RDBMS.
So one transaction can create thousands of records in a table while we
we are iterating over the same table with `each_batch` method, if that
transaction commits it's changes in our last iteration then this method
yields a relation into our block which references to records more than
what we want. This can be one or two or thousands of records more!