Add google tag manager
This adds the google tag manager scripts. They only get activated
when `google_tag_manager_id` is set in the config.
They got added to analyze the sign in, sign up and trial flows of
GitLab. To minimize the amount of JS we load, we only activate
the google tag manager on the sign in/sign up and the trial flow
pages.