Avoid using non-essential frequently-changing command
The dataloss command has been changed a few time recently, causing
tests to fail that expected the behavior before the change.

The dataloss command is not really required in this test to confirm
when nodes are inconsistent. The rake 'replicas' task is enough for
that. The command was included because it was useful to test that it
performs as expected, but the test loses its value when the
behavior under test changes so frequently.

This also fixes a problem in `PraefectManager.replicated?` that
caused the retry loop to end prematurely.