Add links to SAST Configuration UI
This modifies the Security Configuration page for a project to display
a link to the SAST Configuration UI for eligible projects. Roughly
speaking, that means projects without an existing CI file. Future
iterations will [expand][1] the scope of eligible projects.

This is implemented behind the `sast_configuration_ui` feature flag,
which will be removed or enabled by default in
https://gitlab.com/gitlab-org/gitlab/-/issues/231357.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/231373.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/228959