Introduce bare_project factory with minimal associations
Most of the tests don't need all the instances created by project model
callbacks.
We can skip them in a separate factory and make this factory the
default.

This can help to optimize the time spent on project creation.

The disadvantage of this approach is that most of the project
associations aren't created within the factory and future developers
may encounter something like "project.statistics is nil" error.
A comment has been added to the factories file,
assuming it's the first place to go when a developer sees this error.