Add new trainee feature to reviewer roulette
Currently, trainee maintainers are chosen
at a higher rate for reviewer roulette. But
with the move to use domain experts, this has
created a drop in trainee maintainers getting
reviews outside of their domain expertise.

This change adds an additional table to
the roulette table to show potential
trainees for each category to improve
their chances of being given more reviews for
their trainee issue