Commit author for suggestions is note author
Git differentiates the author and committer so that the person who
writes the change can be acknowledged even though someone else committed
it. Suggestions may be proposed by a different person to who applies the
change. If this information is lost, the provenance of the change can be
hard to discern.