Simplify implementation of AjaxLoadingSpinner
The AjaxLoadingSpinner is only used on our branches page. This
simplifies the implementation to not show the button again after the
delete call is finished, as the row is removed anyhow.

We are also replacing the spinner with a proper GlSpinner and the
fontawesome icons with the real sprite_icons.

Furthermore disabled buttons and buttons that actually open a modal for
protected branches had the same event listeners attached. This is not
necessary, as the event listeners were never called.

The files were moved into the branches sub folder as it is only used on
that page.

The test suite has been simplified as well and instead of mocking out
jQuery ajax, we just test the event handler directly. This is work which
helps us with the jquery-ujs to @rails/ujs migration.