Add new lines property to diff files
The lines property wgets used in both inline and parallel views.
This gets it working slightly by allowing the 2 different
views to get rendered with the same data.
It is however, not fully functional.
Whilst discussions get added to rows, they don't get removed from rows.
There is also a lot of other functionality that is currently not working.

https://gitlab.com/gitlab-org/gitlab/-/issues/236143