Fix permission to modify project-level MR rules
This change fixes issue where modification of project-level MR rules is
locked due to the instance-level settings. The interpretation of the
project-level rule must take into account the instance-level setting
value (i.e. if the instance-level setting is false, maintainer and
administrator should be able to edit the approval rules at project
level).

The project policy are also adjusted to allow finer control of the
following 2 project-level settings:

- Allow overriding approvers per merge request settings
- Edit approval rules (add/update)