Add rake task to discover unused partials
This adds a new rake task `discover_unused_partials`. The tool doesn't
handle ee/ well, so we are just scanning our app/ directory for now.

In order to show how to ignore a few false positives values, we ignore
the kaminari partials. Once this works properly, we hopefully can make
it part of static_analysis.