Merge CE and EE welcome page
To not let these two pages diverge (which already happened) we
want to merge them and only embedd the parts from EE that are
necessery.

For self managed, we also remove the question if the user sets
up the account for their company, since this only makes sense
for Gitlab.com where it is more likely that someone starts a group
and is not already using the GitLab instance soley for their
company.