Copy designs to new issue when issue is moved
Adds a set of services and workers to allow an asynchronous copy of
designs from one issue to another.

It happens asynchronously from the in-request issue move as the copy
can take around 15s for 50 designs.

A copy_state state machine has been added to DesignCollection in order
to track the progress of the copy. It is persisted in redis as
DesignCollection is a PORO and is not persisted in PostgreSQL.

The copy involves copying designs, versions and discussions on designs.

https://gitlab.com/gitlab-org/gitlab/-/issues/13426