Fix merge pre-receive errors when load balancing in use
When a user merges a merge request, the following sequence happens:

1. Sidekiq: MergeService runs and updates the merge request in a locked
   state.
2. Gitaly: The UserMergeBranch RPC runs.
3. Gitaly (gitaly-ruby): This RPC calls the pre-receive hook.
4. Rails: This hook makes an API request to /api/v4/internal/allowed.
5. Rails: This API check makes a SQL query for locked merge requests
   with a matching SHA.

Since steps 1 and 5 will happen on different database sessions,
replication lag could erroneously cause step 5 to report no matching
merge requests. To avoid this, we have a few options:

1. Wrap step 5 in a transaction. The EE load balancer will always
   direct these queries to the primary.
2. Use the load balancing sticking mechanism to use the primary until
   the secondaries have caught up to the right write location.
3. Always force the load balancer session to use the primary for this
   query.

Option 1 isn't great because on GitLab.com, this query can take 500 ms
to run, and long-running, read transactions are not desirable.

Option 2 would be nice, but we'd have to save the primary write location
everywhere we lock a merge request or update the
`in_progress_merge_commit_sha`. We do this in multiple places now
(e.g. for fast-forward merges).

We go with option 3 because it is simpler and guarantees that we will
always have a consistent read. However, none of these queries will ever
be routed to a secondary.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/247857