Reflect design collection copy state on frontend
This adds support for seeing the design collection copy state. When an
issue is moved, the designs are moved asynchronously. The user will
initially see a message that their designs are still being copied. If
there is an error during the copy, they will see an error message
displayed but will still be able to upload their designs.

https://gitlab.com/gitlab-org/gitlab/-/issues/13426