Remove negative-margin-top class
Users can click on "Run Pipeline" to start their pipelines on
merge requests.

The layout was achieved by repurposing the "Run pipeline" button
so it shows in the right place in mobile and desktop by adding
some negative margins.

Now this is done by adding a slot to place the button inside the table.